Split cli and lib files
Importing this library into a project with bundler (webpack) needs to
use shebang-loader in order to remove shebang for node env.
