+++ lib.js
@@ -1,14 +1,8 @@
-#! /usr/bin/env node
-'use strict'
-
-const minimist = require('minimist')
+const pump = require('pump')
+const split = require('split2')
 const Writable = require('readable-stream').Writable
 const { Client } = require('@elastic/elasticsearch')
 const Parse = require('fast-json-parse')
-const split = require('split2')
-const pump = require('pump')
-const fs = require('fs')
-const path = require('path')
 const toEcs = require('pino-to-ecs')
 
 function pinoElasticSearch (opts) {
@@ -127,33 +121,3 @@ function pinoElasticSearch (opts) {
 }
 
 module.exports = pinoElasticSearch
-
-function start (opts) {
-  if (opts.help) {
-    console.log(fs.readFileSync(path.join(__dirname, './usage.txt'), 'utf8'))
-    return
-  }
-
-  if (opts.version) {
-    console.log('pino-elasticsearch', require('./package.json').version)
-    return
-  }
-
-  pump(process.stdin, pinoElasticSearch(opts))
-}
-
-if (require.main === module) {
-  start(minimist(process.argv.slice(2), {
-    alias: {
-      version: 'v',
-      help: 'h',
-      node: 'n',
-      index: 'i',
-      'bulk-size': 'b',
-      'trace-level': 'l'
-    },
-    default: {
-      node: 'http://localhost:9200'
-    }
-  }))
-}
